+++ packages/react-editor/src/renderer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Editor, Block, BlockViewRendererPack } from '@chameleon/core';
+import { Editor, Block, BlockViewRendererPack } from '@chamaeleon/core';
 import { useBlock } from './use-block';
 
 type ChildrenRendererProps = {