+++ packages/react-editor/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @chameleon/react-editor
+# @chamaeleon/react-editor
 
 ## 0.0.1
 
@@ -6,4 +6,4 @@
 
 - Initial public release
 - Updated dependencies
-  - @chameleon/core@0.0.1
+  - @chamaeleon/core@0.0.1